[Templates] Update templates
- Try to make use of c++17 features (and disable c++14 support)
- Use implicit false in nullptr checks
- Use const and const& where possible with auto
- Reorder includes from specific to general to avoid having include
  order dependent compilation success. Since part of the includes is
  built in python this is not perfect yet.
- Do some small formatting work to make the generated output more
  readable even when not using clang-format
- Replace magic numbers by static constants

+++ include/podio/CollectionBase.h
@@ -53,7 +53,7 @@ namespace podio {
     virtual bool isValid() const = 0;
 
     /// number of elements in the collection
-    virtual int size() const = 0;
+    virtual size_t size() const = 0;
 
     /// fully qualified type name of elements - with namespace
     virtual std::string getValueTypeName() const = 0;
